+++ packages/react/src/Components/index.ts
@@ -1,4 +1,5 @@
-export { CollectionSidebar } from "./CollectionSidebar";
+export { AppWrapper } from "./AppWrapper";
+export { CollectionsWrapper } from "./CollectionsWrapper";
 export { DataTable } from "./DataTable";
 export { PageHeader, PageTitle } from "./Header";
 export { Pagination } from "./Pagination";
