--- packages/react/src/Pages/Create/index.tsx
@@ -1,10 +1,0 @@
-import { PageTitle } from "../../Components";
-
-export function CreateCollectionPage() {
-  return (
-    <>
-      <PageTitle>Create</PageTitle>
-      <div className="w-full py-2 xl:py-5 max-w-4xl mx-auto" />
-    </>
-  );
-}
