--- packages/react/src/Pages/List/Table/index.ts
@@ -1,2 +1,0 @@
-export { CollectionTable, Pagination } from "./CollectionTable";
-export { SearchField } from "./SearchField";
