--- packages/react/src/Pages/List/index.ts
@@ -1,2 +1,0 @@
-export { CollectionViewPage } from "./CollectionView";
-export { CollectionTable, Pagination, SearchField } from "./Table";
